Use Docker `build()` API with `decode=True`
It seems that Docker `build()` API does not split multiple JSONs by
default, which leads to a Python JSON parser error on some platforms:

  > ./gsc build --insecure-args python test/generic.manifest
  Traceback (most recent call last):
    File "gsc.py", line 46, in build_docker_image
      json_output = json.loads(chunk.decode(encoding))
    File "[...]/python3.9/json/__init__.py", line 346, in loads
      return _default_decoder.decode(s)
    File "[...]/python3.9/json/decoder.py", line 340, in decode
      raise JSONDecodeError("Extra data", s, end)
  json.decoder.JSONDecodeError: Extra data: line 2 column 1 (char 55)

Docker `build()` API has an option `decode=True` to request a JSON
decode through Docker, which seems to split the incoming stream into
separate JSONs, which are fed to the Python JSON parser separately. With
this option, the error goes away. This error was detected on MacOS.

Signed-off-by: Ian Becker <ibecker@cyberreliant.com>