Use Docker API's `decode=True` instead of manual decoding
There was an error during creation of the JSON chunk object which resulted
in the concatenation of two separate JSON blobs. Due to this, the JSON chunk
could not be parsed by the Python JSON parser:

> ./gsc build --insecure-args python test/generic.manifest
Traceback (most recent call last):
  File "gsc.py", line 46, in build_docker_image
    json_output = json.loads(chunk.decode(encoding))
  File "[...]/python3.9/json/__init__.py", line 346, in loads
    return _default_decoder.decode(s)
  File "[...]/python3.9/json/decoder.py", line 340, in decode
    raise JSONDecodeError("Extra data", s, end)
json.decoder.JSONDecodeError: Extra data: line 2 column 1 (char 55)

Docker API has an option to request a JSON decode directly through Docker,
which is able to parse the JSON without the above issue. This error was
detected on MacOS.

Signed-off-by: Ian Becker <ibecker@cyberreliant.com>