Allow to skip SGX token generation
SGX Launch Token (aka EINITTOKEN) is generated right-before Gramine
launch via the `gramine-sgx-get-token` tool. For DCAP platforms, the
SGX token is unused, so there is no need to generate it. This commit
introduces a GSC switch `--skip-token-generation` for this.

This is useful for DCAP platforms (one less file in the Docker
container), and enables such scenarios as `docker run --read-only`.
Also, skipping running `gramine-sgx-get-token` paves a path to removing
Python as a run-time dependency for GSC-built Docker images.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>