Set `PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python` during GSC steps
Core Gramine uses a pre-generated `_pb2.py` file, which became out of
date with the new release of the Google Protobuf Python lib. GSC started
to failing because of this during the build & sign steps.

Technically, core Gramine should not use the pre-generated file but
instead create this file on demand. But this requires some modifications
in core Gramine. A simple workaround for this issue is suggested by
Protobuf devs: set `PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>