[Examples] Fix OpenVINO to build and run
OpenVINO native build got affected by Google Protobuf Python lib update
and requires the `PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python`
workaround. Also, OpenVINO now has a proper ENTRYPOINT, with an absolute
path to the executable (previously it was a relative path to the symlink
to the executable, which confused Gramine and led to failures).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>