Introduce `binary` and `binary_basename` in GSC templates
Previously, GSC templates used `which {{binary}}` to find the absolute
path to the entrypoint binary. GSC sets the `binary` variable here to
the basename of the Docker image's ENTRYPOINT. So the `which {{binary}}`
command worked well for binaries found in PATH, but failed for binaries
not under PATH but rather specified using absolute/relative path.

This commit changes GSC to set the `binary` variable to the full path.
For the cases when the basename of the binary is needed, we introduce
a new variable `binary_basename`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>