Move from `toml` library to `tomli` + `tomli-w`
`toml` Python library (for parsing and generating TOML files) is buggy
and doesn't support the full spec of TOML 1.0.0. This commit replaces it
with more robust `tomli` (for parsing) and `tomli-w` (for generating).

Note that GSC supports both newer versions of Gramine (that also use
`tomli`) and older versions of Gramine (that use `toml`).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>