Add `exec` command to `apploader.sh` entrypoint
Previously, the `apploader.sh` entrypoint bash script spawned Gramine as
a child process (i.e. a simple `gramine-sgx app ...`). This resulted in
the Gramine process not catching the SIGTERM signal that was e.g. sent
via `docker stop`. That's because in a bash script, signals are not
delivered to child processes.

This commit fixes this problem by replacing the bash-script process with
the Gramine process, instead of spawning a child.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>