Concat or override duplicate key values
All envvars are overridden (if duplicates found) in the following
order: First user-provided manifest, if not then GSC-internal manifest,
if not then from original Docker image env.

The only exceptions are: LD_LIBRARY_PATH, PATH, LD_PRELOAD. They are
concatenated instead of overridden. The order of concatenation is same
as above.

Signed-off-by: jkr0103 <jitender.kumar@intel.com>