Fix name of Docker image generated via `build-gramine`
User must provide the output Docker image name in `build-gramine <name>`
command but the generated image prepends `gsc-` which leads to a
confusion about what image name should be specified in config file.

With this commit, the generated Docker image has the name exactly as
specified in the `build-gramine` argument (without `gsc-` prepended).

Signed-off-by: jkr0103 <jitender.kumar@intel.com>