fixup! Concat or override duplicate key values
TODO: use below commit message:

Concatenate or override duplicate envvars

With this commit, duplicate manifest keys (in particular, environment
variables) are overridden in the following order: first from a
user-provided manifest, then from the GSC-internal manifest, and then
from the manifest-shaped environment of the original Docker image.

The only exceptions are PATH, LD_LIBRARY_PATH, LD_PRELOAD envvars which
are concatenated instead of overridden. The order of concatenation is
the same as above.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>