Deprecate `GSC_PAL` and use instead `GRAMINE_MODE`
Previously, to run `gramine-direct`, one had to specify `docker run ...
--env GSC_PAL=Linux`. This was cumbersome because (1) Gramine users
don't need to know the meaning of word "PAL", (2) the value "Linux"
doesn't correspond to known-to-users `gramine-direct`, (3) it requires
special logic in apploader code.

This commit introduces instead `GRAMINE_MODE` envvar with easier
semantics: the value is the one of `direct` and `sgx`, i.e. the mode in
which user wants to invoke Gramine.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>