Deprecate `GSC_PAL` and use instead `GSC_GRAMINE_BINARY`
Previously, to run `gramine-direct`, one had to specify `docker run ...
--env GSC_PAL=Linux`. This was cumbersome because (1) Gramine users
don't need to know the meaning of word "PAL", (2) the value "Linux"
doesn't correspond to known-to-users `gramine-direct`, (3) it requires
special logic in apploader code.

This commit introduces instead `GSC_GRAMINE_BINARY` envvar with easier
semantics: the value is the binary which user wants to invoke.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>